fix: explicit dtype cast in norm() to prevent FlashAttention fp32 crashes

+++ nanochat/gpt.py
@@ -362,7 +362,7 @@ RESEARCH_ALLOWED_KEYS = {
 
 
 def norm(x):
-    return F.rms_norm(x, (x.size(-1),)) # note that this will run in bf16, seems ok
+    return F.rms_norm(x, (x.size(-1),)).to(x.dtype)
 
 class Linear(nn.Linear):
     """nn.Linear that casts weights to match input dtype in forward.
